characterContext.js was splitted into multiple files and moved to src/context/character/
