+++ src/context/character/types.js
@@ -1,0 +1,5 @@
+export const NAME = 'NAME';
+export const QUOTE = 'QUOTE';
+export const COLOR = 'COLOR';
+export const BACKGROUND_COLOR = 'BACKGROUND_COLOR';
+export const CLEAR = 'CLEAR';
